controller

+++ SysAlmacen/src/main/java/pe/edu/upeu/sysalmacen/servicio/MarcaService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface MarcaService {
     public List<Marca> findAll();
-    public Marca findById(long id);
+    public Marca findById(Integer id);
     public Marca save(Marca marca);
     public void delete(Marca marca);
